move sql to separate file
